Remove use of Tany from parameter type checking
Summary:
Instead of using `Tany` to denote the absence of a parameter type, use an optional type where possible.

When we are forced to generate a type (e.g. when checking the body of a function, and an error has been reported already for missing type), generate an error type variable that behaves contravariantly. This forces it to be solved (eventually) to its upper bound, i.e. `mixed`, which is more reasonable a choice than `nothing` for a parameter type.

Differential Revision: D42270619

fbshipit-source-id: 4ac1b33de515d9f1062f4f1326c04a79b5c5a47f