Interpret types with generic classes with no arguments supplied as representing all possiblilities
Summary: Using a class without its generics is still an error in source code, but this will allow the type checker internally to generate them. This was almost working in the source already, but was missing a few cases around dynamic, and also allowing `C <: C<int>` which shouldn't be.

Reviewed By: losvald

Differential Revision: D41947825

fbshipit-source-id: 0316f28bbc2886067e89b7509412fd4f23ff81ff