use inproc hh_fanout_build for hhdg
Summary:
This diff does a little extra plumbing to create a DepGraphWriter in hh_distc and write an hhdg output file. It's fairly fast, using channels to process DepGraphDeltas on a background thread and Rayon to merge into a single file.

I'm adding an iterator over the DepGraphDelta rdeps so we can write the hhdg
file in batches of dependencies, rather than each edge pair one by one.

Reviewed By: edwinsmith

Differential Revision: D42458226

fbshipit-source-id: 8c2073e3a7462b649b086e85fb3650863cee9538