Add missing param type in HHI for pcntl_waitpid()
Summary: `pctnl_waitpid()` takes an "out" parameter which is marked `inout` since Hack does not have a concept of out-only params. The location is unconditionally overwritten by an `int`, but people sometimes like to pass in `null`, so `?int` seems to be a sensible type. Callers will need to handle the nullability after the call.

Reviewed By: jthemphill

Differential Revision: D42398441

fbshipit-source-id: 80a9a5c76c8ecc770d3df7f0dc79fefdd030cd74