Add missing param type in HHI for Serializable::unserialize()
Summary: Generally, we only want to unserialize strings. But at a bare minimum, an implementation should only be considered to be `Serializable` if it can unserialize *at least* strings. That's what the `string` typehint conveys. Since it's an interface and this is only an HHI change, that should be safe.

Reviewed By: jthemphill

Differential Revision: D42398443

fbshipit-source-id: b1c2ace547cba8004be3a040e1ee523b8b66b9c7