use .hhconfig from specified path
Summary:
The design of the re-architecture is that upon start we copy .hhconfig and hh.conf into the state directory, append "--config key=value" overrides, and thereafter we only ever read from there.

This requires that (1) each component take a parameter for which path to find hhconfig/hhconf at, (2) if a component reads hhconfig/hhconf more than once, it should store its hash to make sure it hasn't changed in between. (It won't change in between if the component is being launched by hh in the re-architecture with its temporary local copy, but it might change if the component has been launched directly).

This diff makes that change to hh_decl. It now operates off StartArgs.hhconfig_path.

Reviewed By: zhyty

Differential Revision: D41428765

fbshipit-source-id: 89d00ae7b7a90d6a188570f011ceacb89866c827