fix duplicate RefId handling in make_local()
Summary:
make_local() has a loop that forms groups of objects to download, where
the groups are bucketed by name; all the names are treated the same way
in a conservatively correct manner.

However, the logic to only handle each RefId once (within each symbol
category) inadvertently causes later categories to be skipped, for the same
name. E.g. if the name list includes [X, x, ...], then the existence of a class
named X will cause us to not handle a co-existing function named `x`,
because of the short-circuit logic. (class 'x' finds the RefId from X, then
"continue"s the whole loop.

I ran across this behavior with `class Foo` and `function foo()` in
/test/quick/asm_assert_t.hhas, while testing a change that made functions
case sensitive.

Theoretically this could have affected toplevel constants or modules,
(which are case sensitive), but make_local doesn't do anything special
with those.

The fix is to only skip the set insertion, not continue the whole loop.

Reviewed By: ricklavoie

Differential Revision: D42501865

fbshipit-source-id: 6e3eec46cb3aac72bccaf4ace6164143ac523264