Remove shallow_class_decl (linearized member lookup)
Summary:
In the shallow_class_decl experiment, we hypothesized that we could improve the performance of lazy declaration (i.e., looking up a class member during typechecking when that class has not yet been declared and stored in shared memory) by traversing a linearized list of that class' ancestors and examining their shallow decls (rather than eagerly computing the folded decl for that class by recursively computing the ancestor folded classes). In cases where the member can be discovered early in that traversal (e.g., the member is directly declared in the class in question, or appears early in a list of thousands of ancestors), we can avoid spending time parsing, declaring, serializing, and compressing as many as 4,000 ancestor classes.

We found that it did not perform better. One reason was that we needed two linearizations in order to match the behavior of folded decl (which folded some information using one ordering, and other information using a second ordering), and this resulted in a lot more overhead (we couldn't unify on a single linearization without widespread modifications to WWW, which we never embarked upon). Another reason may have been overfetching: to find the type of a member which was inherited from the end of a long linearization, we could end up deserializing thousands of shallow classes (including the types of all of their members), whereas folded decl would deserialize only the single member we were looking up.

Since performance was still within 25% of folded decl, and memory usage was much better (since there was no need to store folded classes or member heaps), we kept the shallow implementation around. We speculated that it might be better suited for a sharedmem implementation that supports eviction, since shallow classes can be reparsed with a predictable cost (parse a single file), whereas recomputing a folded class can involve re-parsing and re-folding many ancestors. But once we had an implementation of sharedmem which supported eviction, we found that the cost of recomputing evicted folded decls was not a serious problem.

Furthermore, we found that member lookup with linearizations was very sensitive to the shape of WWW. Changes to core class hierarchies like the Ent framework could result in substantial perf or memory regressions in shallow decl, even when folded decl was minimally affected (or not affected at all). Optimizations like the bloom filters for class members helped reduce the overfetching that caused some of these regressions, but I still don't have confidence that shallow decl is as resilient as folded decl in this regard.

Since linearized member lookup is not necessary to support eviction, is awkward to explain and modify, doesn't match the behavior of other representations of inheritance (e.g., in HHVM and HHBC), doesn't perform better than eager folding, and would need to be ported to Rust for type-inference-driven-compilation (whereas a Rust implementation of folded decls exists), we should just delete it.

This diff attempts to remove all of the shallow_class_decl logic, including computation and traversal of linearizations, the Lazy representation of folded classes in Typing_classes_heap, the test suite for the linearizer, and the test targets for typechecking with the shallow_class_decl setting. The shallow_class_decl setting itself will be removed in a follow up diff.

Reviewed By: CatherineGasnier

Differential Revision: D42462059

fbshipit-source-id: 334a46f4d5c3f4728f52d18f4a35701c22a1cc29