Do not create translations at dsts of impossible TransCFG arcs
Summary:
TransCFG may contain impossible arcs between two nodes, if there is a possible
arc between two nodes that represent the same bytecode range, but different
type pre/post conditions.

regionizeFunc() would then create a region that does not contain this
impossible arc due to usage of the region_prune_arcs() logic. This would cause
the arc to not be covered and regionizeFunc() would then try to create another
region starting at the dst of this arc. The translation that would result from
this would be never used.

Fix this by also removing impossible edges from the original TransCFG. This is
a bit tricky because we don't know all possible entry points of this TransCFG.
The obvious entry points are the first node, nodes that have no input edges and
successors of nodes at which we forcibly break regions.

We make an assumption that the profiling provided reasonably good coverage and
that it is unlikely that we could start with arbitrary types at a node that has
input edges that do not support these types.

However, in the unlikely case that the iteration still did not cover all nodes,
we will keep marking them as possible entry points and continue iterating. If
profiling produced these nodes, they must be obviously reachable.

Only once all nodes are covered we remove impossible edges from TransCFG.

Reviewed By: ottoni

Differential Revision: D42496094

fbshipit-source-id: d7330281b52be7ef8062f142a69963a1320f76a9