Handle rewriting multi-return instrs
Summary:
When performing a rewrite if the instruction being rewritten is multi-return and is followed by Select instructions then the rewriter needs to know how to route the Select InstrIds.

This change
(a) detects if the Selects were not re-routed properly.
(b) automatically reroutes them in simple cases.
(c) allows more complex cases.

Reviewed By: edwinsmith

Differential Revision: D42441069

fbshipit-source-id: 03520b409b1112505f33e469dcbe3dcb16330596