Make CoreAllocator stateless
Summary:
`CoreCachedSharedPtr` currently wastes about half of the memory it allocates, due to implementation details.

The control block of `shared_ptr` under libstdc++ takes
- 8 bytes for the vtable
- 8 bytes for the ref counts
- 8 bytes for the provided allocator (because it has state, one pointer)
- 16 bytes for the colocated `shared_ptr`

That is, 40 bytes. The next `CoreAllocator` size class is 64 bytes, so we're wasting 24 bytes due to internal fragmentation. We could add an ad-hoc size class, but note that the 8 bytes for the allocator are also wasted: we only need the special allocator when allocating the control blocks, and then we don't need to carry it around.

This diff changes `CoreAllocator` to be stateless: allocation can only be done while holding a guard which passes the core index through a thread-local variable. This brings the allocation size to 32 bytes, effectively halving memory usage.

In fact, the new API is much more generic, with `coreMalloc()` and `coreFree()` free functions, on top of which `CoreAllocator` is implemented. If a stateful C++ allocator class is needed, it can be implemented on top of those as well, but since the current only use is in `CoreCachedSharedPtr`, we only need a stateless one for now.

An additional benefit of the new API is that we can move most of the implementation to the cpp file, as it is not dependent on `T` anymore.

We'll still need a custom size class if we ever support libc++, as that uses 16 bytes for the refcounts, but we can do that if/when we evaluate libc++.

Reviewed By: luciang

Differential Revision: D41531576

fbshipit-source-id: 04e14cb30b231db3aa196c3ab19496038e229635