Add type hints to `inout` parameters named "$errcode", "$errorcode" and "$success"
Summary:
The PHP standard library often uses `inout` parameters to report the success of a function, independently of the function's return value.

`$errcode` and `$errorcode` refer to integers which describe the kind of error that may have occurred. `$success` refers to a boolean which is true iff the function succeeded. These are all nullable because people will often pass `null` in, and then expect their variable to change to an `int` or `bool` when the function returns.

Reviewed By: periodic1236

Differential Revision: D42463206

fbshipit-source-id: 2e636e96adc072a027fee5f610e44bea16dcaa43