Do not shortcut lambda checking in dynamic mode for expression trees
Summary: In dynamic-check mode, we don't recheck lambdas but instead just assume that any lambda has type `dynamic`. Unfortunately, the expression-tree logic relies on lambda inference to determine the `TInfer` explicit generic parameter to expression trees, and inferring `dynamic` produces errors when the generic parameter on an expression tree is constrained. Solution is to use full checking for lambdas when in an expression tree even in dynamic-check mode.

Differential Revision: D42545538

fbshipit-source-id: 313dd67ba08705a2f1cdca6ec3bff54e13f15f48