Avoid repeat check of lambda body if all parameters are enforced
Summary:
To check that a function, method or lambda is acceptable to be marked `<<__SupportDynamicType>>`, we avoid the second dynamic check of the body (which we'll call the SDT check for short) by looking at explicit parameter and return types. If the parameter types are all fully enforced, *and* the return type is a subtype of dynamic (under dynamic-aware subyping), we accept the function without an SDT check of the body.

But there are circumstances when we can do better.
1. For a lambda without an explicit return type, we can look at the inferred type of the body, and if it is a subtype of dynamic, avoid the SDT check.
2. For method or function with a non-dynamic return type, it can be accepted if each of the types of the actual return values are subtypes of dynamic. For example, `<<__SupportDynamicType>> function foo():I { return new C(); }` is acceptable for `<<__SupportDynamicType class C implements I` even if `I` does not support dynamic.

An easy way to check (2) is to simply wrap the declared return type in `supportdyn` when checking the body. So in the example above, set the return type to `supportdyn<I>`. It should be possible to play the same trick for lambdas e.g. `supportdyn<#1>` where `#1` is a type variable standing for the inferred return type, but this seems to interfere with subtyping on expected types, so instead we simply check the inferred type against `dynamic`.

Note the regression of one of our tests (`bidirectional.good.php`). This is likely due to `supportdyn` not propagating through the expected type to the function type nested there (e.g. `supportdyn<vec<(function(...))>>`). This will be addressed in a future diff.

Reviewed By: vassilmladenov

Differential Revision: D41343659

fbshipit-source-id: 0a089481e609104659fa21273054ae28f468c432