Use ouroboros to delete `DepGraphOpener` and associated classes
Summary:
A `DepGraph` is a Rust object that points into storage in a memory-mapped file.
For ownership reasons, users must create a `DepGraphOpener`, which owns the mmap
data, then call `open()` on it, borrowing that object.

This two object design was forced by Rust's ownership madel, and was always mildly
annoying to use. But it was especially painful when you want a global variable.
Two workarounds were being used:

- `BaseDepGraph`. This classed used the `ouroboros` crate to contain a self-reference
  to both the `DepGraphOpener` and the `DepGraph`, wrapping everything up in one object.

- `UnsafeDepGraph`. Much like `BaseDepGraph`, but used `unsafe`, `std::mem::transmute`,
  and warning comments to achieve the same goal.

This diff does five things:

- Move the `ouroboros` idea into the core `DepGraph` code.
- Delete `DepGraphOpener`
- Delete `BaseDepGraph`
- Delete `UnsafeDepGraph`
- Clean up code to make Rust happy again

Reviewed By: edwinsmith

Differential Revision: D42520569

fbshipit-source-id: 0d5fc9258b0cf68e6f17142e44a18cd8b8d033e6