Fix the base error type on array access on incompatible base type
Summary: We were creating a frash type variable to when `exp[ix]` has an `exp` with a type that doesn't support indexing. This gets solved to nothing and we consequently generate a hole from `nothing` to a `KeyedContainer`. We would like a hole from the existing type (e.g., `nonnull`) to `KeyedContainer` instead.

Differential Revision: D42579047

fbshipit-source-id: 724a720ca9fe739bc0d5a9e78c45ad7ce1d66892