Make dead UNSAFE_CAST removing codemod omit parantheses when not needed
Summary:
We were unconditionally emitting parantheses while removing a dead UNSAFE_CAST in case of capture (say because of a binary expression having a different precedence than its surrounding expression).

This is not possible in many cases such as method calls, variables, etc. This diff implements a helper function to conservatively check if a variable can be captured and use it to omit parantheses during patch generation.

Reviewed By: periodic1236

Differential Revision: D42635739

fbshipit-source-id: e239556cadfd8fc0c1d646e86652f6460386a6de