Generate nothing type instead of fresh type variable for missing class errors
Summary:
The diff D42346015 (https://github.com/facebook/hhvm/commit/9b661d80c46c15f2a3c431866a17cd9bf34b2baf) introduced a regression in hh when run in RE mode. The underlying cause was the following:

(1) Instead of generating `Terr` from a call to `get_class` that returns None, we generate a fresh "error type variable". Usually this is infrequent (only for actually bad class references e.g. `new ClassDoesNotExist()`), but for `hh_distc`, initial checks of a file can produce `None` from `get_class` very frequently, if the class has not yet been loaded. Each such access will produce a distinct type variable.
(2) In themselves, fresh type variables aren't a problem, but in combination with complex control flow they end up getting unioned (e.g. when inserting into a dictionary), and these unions can get very large.

For the specific case of missing classes, it suffices to generate `nothing` instead of a fresh type variable. If a bad reference such as `new ClassDoesNotExist()` is executed, it will fatal, and so it's sound to assume that the expression has type `nothing` even if the error is silenced by `HH_FIXME`. Historically, we supported PHP classes that were "unknown" to the Hack type checker, so fataling might not happen, and the use of `nothing` would be unsound. But this is no longer the case.

Reviewed By: edwinsmith

Differential Revision: D42604440

fbshipit-source-id: 291783165b60eabc8d8449363cb1cfd115d433b0