Do not use TAny for Omitted expression
Summary:
The AST node `Omitted` is used for two purposes right now.
(1) For missing patterns in a `list` destructuring construct e.g. `list(,$x) = foo();`
(2) For missing constant expressions in hhi files e.g. `const vec<int> $vi;`
For (1), an l-value, it makes sense to use the type `mixed` because we want to be able to assign anything to the "absent" pattern. For (2), an r-value, it makes sense to use the type `nothing`, as we do for other "invalid" expressions, as we want to be able to assign it to anything.

Reviewed By: vsiles

Differential Revision: D42640885

fbshipit-source-id: a8412d8a37887636e7bd4ec7ed295284f8b479bc