rebuild fixtures
Summary:
The new mstch go generator introduced in D32620306 (https://github.com/facebook/hhvm/commit/692c369b55d0e51754cb78a86a073c5459554494) has a problem: it is not hermetically sealed, and it generates files that contain the full filesystem path of where the source .thrift files are checked out to on disk.

We can see this in diffs to the fixtures, starting in D41855177 (https://github.com/facebook/hhvm/commit/2429bf0fb345e3f6d4ff80a1fb816293688d1f8a). But, for example: D42583818 (https://github.com/facebook/hhvm/commit/b5d79250c1bf30a2c32dcf6ba8731ba3f6bc2609) touches every single .go generated file for no reason - except to modify the full filesystem paths that get included.

This is not permissible. As a requirement, the generator must only generate files which are agnostic to the checkout path, and that would be invariant and unchanging whenever the checkout path changes. A requirement is that build artifacts are cacheable, and files that contain full filesystem paths are not cacheable.

In this diff, we remove the full filesystem paths and merely replace them with placeholders. As future work, the placeholders may be replaced with corrected, invariant file paths.

Reviewed By: echistyakov

Differential Revision: D42667605

fbshipit-source-id: 213a19de0eff8bc3703757d4d6454301aad8cf78