asm_volatile_pause: use isb on aarch64 (#1904)
Summary:
On aarch64, we've found ```isb``` to be a better analogue to the x86 ```pause``` instruction. The ```SpinLock.TryLock``` test/microbenchmark shows a one to two order-of-magnitude reduction in runtime (its runtime is highly variable with and without this change due to its very race-y nature).

Testing:
I ran this on aarch64 with Ubuntu 22. Sample runtimes of ```SpinLock.TryLock``` before:
```
[       OK ] SpinLock.TryLock (50903 ms)
```
```
[       OK ] SpinLock.TryLock (32773 ms)
```
```
[       OK ] SpinLock.TryLock (70178 ms)
```
and after:
```
[       OK ] SpinLock.TryLock (728 ms)
```
```
[       OK ] SpinLock.TryLock (2003 ms)
```
```
[       OK ] SpinLock.TryLock (2 ms)
```

X-link: https://github.com/facebook/folly/pull/1904

Reviewed By: yfeldblum

Differential Revision: D42329777

Pulled By: Orvid

fbshipit-source-id: 66f4b88f51f1395a5fa53aaa2eb5ee5ef916e74f