Fix regression in pessimisation errors due to like types in virtualized types
Summary:
When typing expression trees, we artificially infer a Hack type for the virtualized type expression, and feed that back into as an "explicit" type argument to `makeTree`. This is not the usual way that type inference works, and for pessimised code we end up using a like type (or something more complex, such as `supportdyn<~t>`) for the explicit type argument. If there is a constraint on the explicit type parameter `TInfer` then the constraint will not be satisfied. We really want to infer something without a like type, and let the parameter passing to `makeTree` handle likes.

The fix is to crudely strip off likes, looking underneath unions and `supportdyn`.

Differential Revision: D42637428

fbshipit-source-id: f804a791e22a7e6c6c1ecdaba68c3b89e315638e