Implement FuncFamily::StaticInfo calculation using extern-worker
Summary:
StaticInfo is the information common to all methods in a FuncFamily
which doesn't change (hence "static"). It's used during analysis to
avoid iterating over potentially large FuncFamily lists. Move the
calculation of such StaticInfos into extern-worker.

For the most part this is straight-forward. The StaticInfo can exist
on the FuncFamily and the FuncFamily logic can build them and combine
them together along with the rest of the state. There is, however, one
complication. In order to initially calculate the StaticInfo, you need
the actual methods (it cannot be derived from the MethTabEntry). We
could supply the actual methods to the job, but this can be expensive,
and we don't need very much. Instead when we're building the method
tables, store the relevant information as part of the methodFamilies
table. This is already sent to the job, so it can be used to jump
start the StaticInfo calculation.

This was the last remaining use of the methods table in Index, so that
can now be removed.

Once these changes were made, there was an unexpected memory
regression. It turns out that this was caused by moving allocations
around causing the heap to fragment more. Clearing some containers
earlier resolved this issue.

Reviewed By: jano

Differential Revision: D42692756

fbshipit-source-id: 45c58b59452a9738eb9325871874e91538f8035d