Avoid _generating_ new expression during elaboration
Summary:
During `naming` we currently _generate_ a new, invalid expression for class variables in hhis. This is done only to satisfy a NAST check (`nastInitCheck`) which ensures that all class vars are initialized.

We'd like to avoid generating new expressions during elaboration since it requires us to make concrete the type annotations on the AST.

This diff adds the corresponding hhi condition to `nastInitCheck` so we no longer need to generate an unused, invalid expression for hhis.

Reviewed By: shayne-fletcher

Differential Revision: D42673993

fbshipit-source-id: c84ff383d32a53658aef0c4f938105c3af92f04d