introduce `Invalid` expression marker
Summary:
Adds an explicit constructor `Invalid of ('ex, 'en) expr option` into `expr_`.

This is intended to replace use of the current invalid expression marker, a lambda which raises an expression in its body, generated during the naming phase. The inner expression is optional since we may need to generate an invalid marker due to the absence of an expression in some cases.

The main reason for the change is the expectation that the naming phase will be put in front of HackC where we must be able to distinguish and handle invalid expressions.

The diff also handles the new expression in typing where we simply map the invalid expression from `(unit,unit) expr` (i.e. a NAST expr) to `(locl_ty,saved_env) expr` (i.e. a TAST expr).

This also has the benefit of retaining the original expression for user presentation.

Reviewed By: Wilfred

Differential Revision: D42633258

fbshipit-source-id: a49ce752a76fc05efac28c89b62380752a18021f