Kill Mac/Windows fsnotify implementations
Summary:
Hack does not run on Mac or Windows, so we don't need this code. The remaining
implementation use inotify to watch the file system.

The Mac implementation used fsevents, which is now dead, and removed in this
diff. The fsevents code used a naked pointer, which is not supported in OCaml 5.

Reviewed By: shayne-fletcher

Differential Revision: D42753662

fbshipit-source-id: 045a88705947f192134ef72e28c919c0224da75e