Speed up jemalloc checks
Summary:
`usingJEMalloc()` and related functions are very hot (called on pretty much every allocation/deallocation path in folly) and the use of `static` to cache the result is suboptimal, as it has an effect on code size (slow path is large, reducing the chance of inlining) and it is more expensive than necessary, because both the guard variable and the value need to be loaded, and they may even be in different cache lines.

We can use a tri-state atomic instead.

See generated code with godbolt: https://godbolt.org/z/GsrKdzPeM
The entirety of the function is now just
```
fast():                               # fast()
        movzx   eax, byte ptr [rip + _ZN14FastStaticBoolIZ4fastvE4InitE5flag_E.0]
        test    al, al
        je      .LBB0_2
        setg    al
        ret
.LBB0_2:
        jmp     FastStaticBool<fast()::Init>::getSlow() # TAILCALL

```

which is very inlineable, and feedback-driven compilation can even move the branch on the **returned value** before the unlikely branch that confirms initialization.

Reviewed By: dmm-fb

Differential Revision: D42609469

fbshipit-source-id: 1f7cbfe96e6e057a1f7b4f97739fac1b057c1078