<<__NoAutoDynamic>> support on type alias
Summary:
This diff adds support for the `<<__NoAutoDynamic>>` attribute on type aliases.

Additionally it updates the naming elaboration for type aliases so that implicit bounds are added to the ast in implicit pessimisation, resulting in

```
<<__NoAutoDynamic>>
class C {}

newtype Bar = C;
```

correctly raising an error.

Implementation detail: only `__Docs` and `__NoAutoDynamic` attributes are allowed on type aliases.  `__Docs` attributes were not stored in the decls; this diff does not modify the direct decl parser handling of `__Docs` attributes.

Reviewed By: andrewjkennedy

Differential Revision: D42603177

fbshipit-source-id: b37a489cfe723657d8a4e54faa9529226d572430