Fix NoAutoDynamic for returns
Summary: When encountering a __NoAutoDynamic attribute, disable implicit pessimisation in the type checker, via the everything_sdt option. It would be better if the type checker didn't look at the flag, and implicit pessimisation was handled entirely in decl and naming, but there are a few places (e.g., function returns) where that requires a lot of refactoring.

Reviewed By: vassilmladenov

Differential Revision: D42807832

fbshipit-source-id: 58c30220b33ea3da2a62827549fbe650cace4c3a