Always cache final results of find_defining_inst_for_remat_enter
Summary:
The find_defining_inst_for_remat algorithm discards results that involve
recursive definitions, because these results are partial. But it also
unnecessarily discarded the final result, if it was part of a cycle. This
caused pathological quadratic behavior with large cycles involving a lot of
phis, as a lookup on each phi would need to traverse the whole cycle.

If we cache the final result, which is safe to cache, it is likely to break the
cycle and the subsequent lookup will be able to cache all other results,
stopping the quadratic behavior.

Reviewed By: ricklavoie

Differential Revision: D42831472

fbshipit-source-id: 8480c715d26584f92928533d254bc8191f129cd0