Move HHBBC interface vtable calculation to use extern-worker
Summary:
Convert HHBBC's interface vtable calculation over to use
extern-worker. The algorithm has two components. The first is
calculating the conflict graph for each interface. This is the set of
interfaces implemented by every class which implements that
interface. This is, by far, the most expensive component of the
algorithm. Luckily it can be easily calculated while building class
subclasses. The second component is using the conflict graph to find
minimum non-conflicting slots. This is quite cheap, but hard to
distribute. We'll still calculate this locally.

Reviewed By: kavoor

Differential Revision: D42838031

fbshipit-source-id: 300f079999368a3f9f259d5d4cccb5b7f9947948