Fix ExecutionScope use after move issue
Summary:
User encounter use-after-move lint warning on D42243397 (https://github.com/facebook/hhvm/commit/8176184e7a8b635a25301ca2badfbe849bbcae43). Fixing the issue in that diff was out of scope, since the warning was for lines that were not modified as a part of the diff.

After looking at the code, I believe this to be a legitimate issue, so I'm fixing it.

Reviewed By: maalbash

Differential Revision: D42819227

fbshipit-source-id: a4b66712a1126a87783b51b9b5eafbda3366d1fd