fix trait/interface sdt hierarchy check
Summary:
The SDT hierarchy checks where performed only on the immediate parents of a class.  As a consequence, when a class `C` inherits an `implements I` constraint via a trait use, the check was not ensuring that if `I` is SDT then `C` is SDT.

This diff ensures that hierarchy checks are performed wrt _all_  parents of the class.

Reviewed By: andrewjkennedy

Differential Revision: D42805579

fbshipit-source-id: 8f19de36353c965cc7b0b748dccd162c485d84e3