`Ctx` shouldn't me mutable
Summary:
The `Ctx` type is used to record contextual information about where an AST node occurs for use in the top-down phase of the elaboration.

Previously, we has a mut ref argument which meant that mutations to the context where propagated through the fused transformations which is what we want.

The problem with this is that we were also carrying the modified context between sibling AST nodes, e.g. if we have some constructor

```
As(Box<Expr>,Box<Hint>)
```

when we transform the `Expr` we were inadvertantly using the modified context when transforming the `Hint`

This diff make the `Ctx` a value type which must implement `Copy`. It also updates our `Pass` composition code so that we continue to propagate the context through fused transforms but we no longer propagate it across sibling AST nodes

Reviewed By: rjbailey

Differential Revision: D42888937

fbshipit-source-id: 22ccb68fb3b68419ffb368e728ad4e617fcb122d