Eliminate use of `Typing_error.Primary.t` from elab passes
Summary:
There are a handful of uses of _typing_ errors in the naming phase. The use there is because those errors aren't really typing errors in the sense that they do not require a typing environment to determine there is an error.

As well as being 'wrong', it causes problems for us since oxidize can't handle nested modules.

This diff adds the relevant typing errors to `Naming_error.t` but maintains the `User_error.t` reporting behaviour

Reviewed By: shayne-fletcher

Differential Revision: D42927610

fbshipit-source-id: e204ea584c603b17d30291bd8e7006bcbb174f61