Add dune build support
Summary:
Here we add config needed to build ERRPY via dune.

The idea is to have `dune` driving the build, with `cargo` being invoked for the rust. This works ok via `dune build` and we are able to output a library which can be installed and referenced by pyre.

This diff consists of the following core elements:
 - `dune` and `cargo` config files added to drive the build.
 - Changes to the `README.MD` with regards to how to build ERRPY.
 - Formatting of the OCaml code (as `ocamlformat` config is added in this diff)

Note that rust is specified as a dependency in `dune-project` (as `conf-rust:build`) - this should force cargo/rust to be installed [via opam](https://opam.ocaml.org/packages/conf-rust/) on the machine running the dune build.

## Dependencies:
 - The following dependencies have been symlinked in: `ocamlpool, ocamlrep, ocamlrep_derive, ocamlrep_ocamlpool, rust_to_ocaml, utils/hash` under `./vendor`
 - I had to copy over ocamlpool, I was unable to get the build to work when symlinked

Hopefully [ShipIt](https://www.internalfb.com/intern/wiki/Open_Source/Internal_Tools/ShipIt/) will automatically expand the symlink for the OSS distribution of ERRPY. If not we will have to copy over the symlinked versions until they are turned into OSS projects.

## Autocargo

I was not able to repoint the `src/TARGETS` file to use the versions of the dependencies under `./vendor` and have buck2 build ERRPY correctly. Hence `arc autocargo` was not usable to generate the cargo config. Will need to revisit this in the future as it’s more convenient to use autocargo.

Reviewed By: stroxler

Differential Revision: D41897777

fbshipit-source-id: 74845703930498ff11499253ff524bee090d9339