Non-existent classes in HHBBC should result in Bottom, not unresolved classes
Summary:
For historical reasons, if you try to resolve a class which doesn't
actually exist (or is uninstantiable), we produce an "unresolved"
class. All the reasons for this are no longer relevant, so instead
treat such a missing class as Bottom instead. Doing so revealed a
couple of pre-existing bugs and some places which was conflating
unresolved classes with non-existing classes. (A non-existing class
was always unresolved, but an unresolved class was not necessarily
non-existing).

Reviewed By: jano, edwinsmith

Differential Revision: D42877899

fbshipit-source-id: 69180b33e64ecf426b65cf67b54285eaea1c35b3