Fix flaky ParallelConcurrencyController tests
Summary: Both `ParallelConcurrencyControllerTest.DifferentOrdering1` and `ParallelConcurrencyControllerTest.DifferentOrdering2` have been flaky for some time. This is due to a race condition involving the folly::Latch use in both tests. Prior to this change, it was possible to latch.wait() to be satisfied before ParallelConcurrencyController had an opportunity to update its internal counters (which happens in the `onExecuteFinish` callback).

Reviewed By: maalbash

Differential Revision: D42824561

fbshipit-source-id: 6ceafdb2f21a43b1d383aaf808375bc1c99d04d9