Update once_cell-1.15.0 to once_cell-1.17.0
Summary:
D42948747 is adding a dep that requires this update. Because `once_cell`
is so widely used by other third-party crates, this update needs to be
landed quickly to avoid races in `third-party/rust/BUCK`. And because
`once_cell` is so widely used by internal code, this update is going to
schedule a lot of jobs.

Nothing that can break was changed:
https://github.com/matklad/once_cell/blob/v1.17.0/CHANGELOG.md

Reviewed By: Imxset21

Differential Revision: D42950546

fbshipit-source-id: 535711c003d52f4566e7287ac67a422af42db5f3