Distinguish between type/ctx after decl phase
Summary:
HackAST relies on `--hh-type-at-pos-batch`, which dumps type info
via `Typing_print`.  Since contexts become types after decling,
there is currently no way to distinguish `with { type A }` from `with { ctx A }`
(in theory a decl lookup is possible but that's impractical because `Typing_print`
is the very core typechecker with minimal depenencies).

Extend `Typing_defs.class_type_refinement` exact & loose variant to include a boolean flag; it is true if the type represents the context but false otherwise. Use this opportunity to rename types as follow:
* `class_type_refinement` -> `refined_const_bound` (since now it applies to both `type`s and `ctx`s in the refinement)
* `cr_types` -> `cr_consts`

Finally, modify the JSON to include an extra bool "is_ctx", so that HackAST
can properly sugar names that refer to contexts.

___

Other alternatives considered:
1) (V2) least intrusive but blocked by the design of smart constructors / direct-decl parser: in `Typing_defs.class_refinement` add a `SSet` of names that refer to contexts as a sibling to `cr_types` (mapping from names to refined `type`/`ctx` members).
E.g., for `with { type A; ctx B }`, the set will contain "B" but not "A".
2) (V3) Less intrusive, equally (non-)expensive but more odd: duplicate a flag inside `class_type_refinement` (now `refined_const_bound`).

Reviewed By: mpu

Differential Revision: D42826924

fbshipit-source-id: b2f3eb26efe6aa591ac203e3043319be40dc8906